requirements.txt
gradio==3.23.0
torch==1.13.1
torchvision==0.14.1
pillow==9.5.0
torchmetrics==0.11.4
matplotlib==3.7.1
wandb==0.15.2
tqdm==4.65.0
numpy==1.24.2
timm==0.6.13